rename files and fix scripts
